+++ ormar/models/helpers/__init__.py
@@ -1,5 +1,7 @@
 from ormar.models.helpers.models import (
+    check_required_meta_parameters,
     extract_annotations_and_default_vals,
+    meta_field_not_set,
     populate_default_options_values,
 )
 from ormar.models.helpers.pydantic import (
@@ -15,7 +17,9 @@ from ormar.models.helpers.relations import expand_reverse_relationships
 from ormar.models.helpers.sqlalchemy import (
     populate_meta_sqlalchemy_table_if_required,
     populate_meta_tablename_columns_and_pk,
+    sqlalchemy_columns_from_model_fields,
 )
+from ormar.models.helpers.validation import populate_choices_validators
 
 __all__ = [
     "expand_reverse_relationships",
@@ -28,4 +32,8 @@ __all__ = [
     "get_pydantic_field",
     "get_potential_fields",
     "get_pydantic_base_orm_config",
+    "check_required_meta_parameters",
+    "sqlalchemy_columns_from_model_fields",
+    "populate_choices_validators",
+    "meta_field_not_set",
 ]