+++ ormar/queryset/queryset.py
@@ -555,6 +555,7 @@ class QuerySet:
             self.model.extract_related_names()
         )
         updates = {k: v for k, v in kwargs.items() if k in self_fields}
+        updates = self.model.validate_choices(updates)
         updates = self.model.translate_columns_to_aliases(updates)
         if not each and not self.filter_clauses:
             raise QueryDefinitionError(